+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Added
+
+- New ``solve_ls`` function to solve linear Least Squares problems
+
 ## [1.4] - 2020/07/04
 
 ### Added